Automatische Buildnummer (vorbereitet)
git-svn-id: svn://svn.berlios.de/dsa-hl@1235 4c79e8ff-cfd4-0310-af45-a38c79f83013

+++ rl/trunk/engine/core/include/ConfigurationManager.h
@@ -68,7 +68,7 @@ namespace rl
 			Ogre::String getModulesRootDirectory();
 
 			Ogre::String getEngineVersionString();
-			int          getEngineBuildNumber();
+			long          getEngineBuildNumber();
 			const rl::CeGuiString&  getAboutText();
 			
 		protected:
